+++ stdnet/odm/base.py
@@ -387,7 +387,7 @@ class ModelType(type):
         fields = get_fields(bases, attrs)
         new_class = super(ModelType, cls).__new__(cls, name, bases, attrs)
         ModelMeta(new_class, fields, **meta)
-        class_prepared.send(sender=new_class)
+        class_prepared.fire(new_class)
         return new_class
 
     @classmethod