+++ stdnet/utils/encoders.py
@@ -65,13 +65,15 @@ for encoding and decoding data to and from the data server.
         return False
 
     def load_iterable(self, iterable, session=None):
-        '''Load an ``iterable``. By default it returns a generator of
-data loaded via the :meth:`loads` method.
+        '''Load an ``iterable``.
 
-:param iterable: an iterable over data to load.
-:param session: Optional :class:`stdnet.odm.Session`.
-:return: an iterable over decoded data.
-'''
+        By default it returns a generator of data loaded via the
+        :meth:`loads` method.
+
+        :param iterable: an iterable over data to load.
+        :param session: Optional :class:`stdnet.odm.Session`.
+        :return: an iterable over decoded data.
+        '''
         data = []
         load = self.loads
         for v in iterable: